backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
email-validator==2.2.0

# (Postgres driver temporarily omitted due to Python 3.13 wheel availability issues.)
# When ready for Supabase on a stable version (e.g. Python 3.12) add:
# (Tạm thời không cài Postgres driver trên Python 3.13 vì thiếu pg_config và wheel.)
# Khi cần kết nối Supabase có thể:
#   1. Dùng Python 3.12 rồi thêm: psycopg[binary]>=3.1,<3.3
#   2. Hoặc cài PostgreSQL client (để có pg_config) rồi dùng psycopg-c.
# Hiện tại chạy SQLite dev ổn định.

# Optional heavy deps (only needed if you implement real image AI later)
# pillow==11.0.0
# numpy==2.1.3

# Recreate venv with Python 3.10 then uncomment the line below for Supabase (psycopg3 driver)
psycopg[binary]==3.2.1
